// src/instruments/brassEnsemble.d.ts
export type EnsemblePartDef = {
    part_id: string;
    name: string;
    instrument: string;
    staves: number;
};
/**
 * BRASS ENSEMBLE (default)
 * Practical, common church/community brass ensemble:
 * - 2 Trumpets
 * - 1 Horn
 * - 1 Trombone
 * - 1 Bass Trombone (or 2nd Trombone)
 * - 1 Tuba
 */
export declare const BRASS_ENSEMBLE_PARTS: EnsemblePartDef[];
//# sourceMappingURL=brassEnsemble.d.ts.map